Add Supporting files
Just for fun I wanted to
give TypeScript a try for tests.
It really doesn't give me much
value here, but I thought it
was worth figuring out how to
get it set up for Buidler...
it was a huge pain. I had to do a
lot of digging to get things working;
if I knew more about TS, it probably
would have been more intuitive.