Add ERC1155 EQ Token
The ERC1155 token is a token standard
that allows for multiple NFTs and FTs.
This is a newer standard that can
save a significant amount of gas,
apparently.

The app will allow users to stake
any ERC20 token they want, and
over time they will earn EQToken ID 0
which is a fungible token - Platinum.
The stakers can then burn Plat to
purchase available NFTs on the same
EQ Token contract.