Add Token Farm
At this point the way rewards are
distributed is fairly simplistic;
however, I did go with a new approach
to distribution that I haven't seen elsewhere.

== The What:

Rather than mint Platinum to all stakers
in a sort-of PUSH approach, I instead keep
track of the rewards accrued, and I set it up
so users can PULL their portion of the rewards.

== The How:

To do this I keep track of each individual
deposit so I can accurately calculate rewards
for each. I take into account when the deposits
were made, and when / if rewards were already
"paid out" for those deposits.

This is an attempt to avoid unnecessary work
of paying out rewards to inactive users.
As the app grows, there could be many many
inactive users, and this could get expensive.

With this approach, we can freely (at no cost)
calc rewards so the user can see how many
awarded tokens are available to claim. That user
can then decide whether or not to pay the small
gas fee to claim the reward. This puts the
burden of the gas fee on the user, for sure,
but it's very tiny when compared to the large
fee the app maintainer would incur to distribute
rewards to all stakers.

== In Sum:

Advantages:
- gas costs for realizing / claiming the
rewards are shifted to the users, and
the contract owner is spared the potentially
high fees of distributing rewards to
a large staker-base every month.

Disadvantage:
- some users may consider this shift in
costs a bad thing, but this allows the
project to stay viable.