TokenFarm Fix Unbounded Loops
**unbounded** loops are a danger
because the size of the array can
grow to a point where a full iteration
would be so expensive as to make
execution of the function impossible.

Instead, it was recommended that I pass
the function an array of tokens for
which to claim rewards.

Add stakedTokens()

Since we're not just iterating over all
user-staked tokens, we require
a list of staked tokens before we make
the claimRewards() request.